move xpu connector
